+++ projeto-galaxt-tech/src/components/Footer/index.jsx
