+++ projeto-galaxt-tech/src/components/Header/styled.js
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+/* import styled from "styled-components"; */
 
 export const ContainerHeader = styled.header`
   position: fixed;
